[QT-706] Support optional token auth strategy
Unaunthenticated downloads are a nice feature as they don't require
users to specify a token when creating an action. They do come with a
downside of being more likely to be throttled.

This adds support for 2ish auth methods:
* Unaunthenticated
* Token auth using either the github-token input or by setting
  GITHUB_TOKEN in the environment.

As part of this we also added tests for all three strategies.

Signed-off-by: Ryan Cragun <me@ryan.ec>