auth: set auth token on asset download if present
We previously added support for using an auth strategy when looking up a
release asset by the tag, however, we didn't ever thread it through to
our asset download. This change will resolve the auth object and set the
bearer token on the asset download if we've got token auth.

We also fix debug logging in actions by using actions/core logging.

Signed-off-by: Ryan Cragun <me@ryan.ec>